+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 Release notes
 ==============
 
+### Unreleased
+
+* Added a new method `SkipRelationshipIf` to relationships that allows specifying when a relationship will be entirely omitted from the resource (i.e., no relationship links will be present). Use this together with the `Get...Skip` methods. If not (as has always been the case), the links will be present, but `GET` operations against them will return errors if the relationship's getter returns `Skip`.  
+
 ### 0.20.7 (2022-08-18)
 
 * Errors returned for invalid JSON no longer uses the exception message, which means less implementation details are leaked (but the message may be somewhat less useful)